# openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/login": {
      "post": {
        "operationId": "UserController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid username or password"
          }
        },
        "summary": "Login a user",
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "EventsAR API",
    "description": "API for EventsAR application",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "Authorization bearer token"
      }
    },
    "schemas": {
...
      "UserLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "gosho"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "123456"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
    }
  }
}